requirements.txt
python-dotenv>=1.0.1
playwright>=1.45.0
langgraph>=0.2.0
pydantic~=2,<3
anthropic
firecrawl-py

pandas
rich
requests
beautifulsoup4
Pillow>=10.0.0
